refactor(utils) - rearrange utils package
