+++ internal/server/handlers.go
@@ -7,8 +7,10 @@ import (
 
 	"github.com/ronfelsenfeld/go-proxy/internal/config"
 	"github.com/ronfelsenfeld/go-proxy/internal/logger"
-	"github.com/ronfelsenfeld/go-proxy/internal/utils"
+	apiUtils "github.com/ronfelsenfeld/go-proxy/internal/utils/apiutils"
 )
+
+
 
 func pingHandler(responseWriter http.ResponseWriter, request *http.Request) {
 	response := map[string]string{"message": "pong"}
@@ -19,7 +21,7 @@ func pingHandler(responseWriter http.ResponseWriter, request *http.Request) {
 }
 
 func proxyHandler(responseWriter http.ResponseWriter, request *http.Request, configuration *config.Config) {
-	requestBody, err := utils.DecodeJSONBody(request)
+	requestBody, err := apiUtils.DecodeJSONBody(request)
 	if err != nil {
 		logger.Error.Printf("❌ %s", err.Error())
 		http.Error(responseWriter, err.Error(), http.StatusBadRequest)
@@ -29,14 +31,14 @@ func proxyHandler(responseWriter http.ResponseWriter, request *http.Request, con
 	requestBody[configuration.InjectKey] = configuration.InjectValue
 	logger.Info.Printf("🔨 Modified request body: %+v", requestBody)
 
-	modifiedBody, err := utils.EncodeJSONBody(requestBody)
+	modifiedBody, err := apiUtils.EncodeJSONBody(requestBody)
 	if err != nil {
 		logger.Error.Printf("❌ %s", err.Error())
 		http.Error(responseWriter, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	upstreamResponse, err := utils.ForwardRequest(request, modifiedBody, configuration)
+	upstreamResponse, err := apiUtils.ForwardRequest(request, modifiedBody, configuration)
 	if err != nil {
 		logger.Error.Printf("❌ %s", err.Error())
 		http.Error(responseWriter, err.Error(), http.StatusInternalServerError)
@@ -44,7 +46,7 @@ func proxyHandler(responseWriter http.ResponseWriter, request *http.Request, con
 	}
 	defer upstreamResponse.Body.Close()
 
-	utils.CopyHeaders(upstreamResponse.Header, responseWriter)
+	apiUtils.CopyHeaders(upstreamResponse.Header, responseWriter)
 
 	responseWriter.WriteHeader(upstreamResponse.StatusCode)
 	io.Copy(responseWriter, upstreamResponse.Body)
