+++ internal/server/server.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/ronfelsenfeld/go-proxy/internal/config"
 	"github.com/ronfelsenfeld/go-proxy/internal/logger"
-	"github.com/ronfelsenfeld/go-proxy/internal/utils"
+	apiUtils "github.com/ronfelsenfeld/go-proxy/internal/utils/apiutils"
 )
 
 func Router(configuration *config.Config) http.Handler {
@@ -15,8 +15,9 @@ func Router(configuration *config.Config) http.Handler {
 
 	requestRouter.HandleFunc("/proxy", func(responseWriter http.ResponseWriter, request *http.Request) {
 		logger.Info.Println("🔍 Request received:", request.Method, request.URL.Path)
-		
-		if utils.GetIsPostRequest(request) || utils.GetIsPutRequest(request) {
+
+	
+		if apiUtils.GetIsPostRequest(request) || apiUtils.GetIsPutRequest(request) {
 			proxyHandler(responseWriter, request, configuration)
 		} else {
 			logger.Error.Println("⛔️ Method not allowed:", request.Method, request.URL.Path)
